+++ src/HelperUrl.js
@@ -2,51 +2,73 @@ const HelperString = require('./HelperString.js');
 
 class HelperUrl 
 {
-    static getBaseUrl() 
-    {
-        const baseHref = HelperUrl.getBaseHref();
-
-        if (baseHref) {
-            return baseHref;
-        }
-
-        return HelperUrl.getBaseUri();
-    }
-
-    static getBaseUri() 
-    {
-        // Path ends with an "/", it must be a directory
-        if (window.location.pathname.slice(-1) == '/') {
-            return window.location.protocol + '//' + window.location.host + window.location.pathname;
-        }
-        
-        const matchExtension = window.location.pathname.match(/\.[\w]{2,4}$/);
-
-        // There is no extension, it may or may not be a directory
-        // but lets assume it is.
-        if (! matchExtension) {
-            return window.location.protocol + '//' + window.location.host + window.location.pathname + '/';
-        }
-
-        // There is an extension, likely a file, lets remove the last part of the url.
-        return window.location.protocol + '//' + window.location.host + window.location.pathname.replace(/\/[^\/]+$/, '') + '/';
-    }
-
+    /**
+     * Returns the base href of the document, usefull when working with
+     * relative urls.
+     * 
+     * @returns string An url
+     */
     static getBaseHref() 
     {
-        const base = document.querySelector('base');
+        return HelperUrl.getDocumentBaseHref() ?? HelperUrl.getWindowLocationBaseHref();
+    }
 
+    /**
+     * Attempts to figure out the base href based on the current value of 
+     * window.location
+     * 
+     * @returns string An url
+     */
+    static getWindowLocationBaseHref() 
+    {
+        const location = window.location;
+        var href;
+
+        // Path ends with an "/", it must be a directory
+        if (location.pathname.slice(-1) == '/') {
+            href = location.protocol + '//' + location.host + location.pathname;
+        } else {
+            const matchExtension = location.pathname.match(/\.[\w]{2,4}$/);
+
+            // There is no extension, it may or may not be a directory
+            // but lets assume it is.
+            href = !matchExtension
+                ? location.protocol + '//' + location.host + location.pathname
+                // There is an extension, likely a file, lets remove the last part of the url.
+                : location.protocol + '//' + location.host + location.pathname.replace(/\/[^\/]+$/, '');
+        }
+
+        return HelperString.rtrim(href, '/') + '/';
+    }
+
+    /**
+     * If specified, returns the href attribute in the <base> tag,
+     * returns null otherwise.
+     * 
+     * @returns string|null
+     */
+    static getDocumentBaseHref() 
+    {
+        const base = document.querySelector('base');
         if (! base) {
             return null;
         }
 
         const href = base.getAttribute('href');
-
         if (! href) {
-            return href;
+            return null;
         }
 
         return HelperString.rtrim(href, '/') + '/';
+    }
+
+    /**
+     * @param string uri 
+     * @returns bool
+     */
+    static isAbsolute(uri) 
+    {
+        return uri.substr(0, 4) == 'http';
     }
 }