+++ tests/test-index.js
@@ -1,0 +1,5 @@
+window.Http = require('../src/Http.js');
+window.HelperString = require('../src/HelperString.js');
+window.HelperUrl = require('../src/HelperUrl.js');
+window.Convert = require('../src/Convert.js');
+window.CastDown = require('../src/CastDown.js');